requirements.txt
flask==2.2.3
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
werkzeug==2.2.3 